requirements.txt
# requirements.txt

alembic==1.18.1
attrs==25.4.0
babel==2.17.0
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dbfread==2.0.7
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
et_xmlfile==2.0.0
flasgger==0.9.7.1
Flask==3.1.2
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.0.5
flask-talisman==1.1.0
Flask-WTF==1.2.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
mistune==3.2.0
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
psycopg2==2.9.11
pycparser==3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==1.4.48
typing_extensions==4.15.0
tzdata==2025.3
ua-parser==1.0.1
ua-parser-builtins==202603
urllib3==2.6.3
user-agents==2.2.0
Werkzeug==3.1.5
wsproto==1.3.2
WTForms==2.3.3
